requirements.txt
alembic==1.2.1
cffi==1.13.2
Click==7.0
cryptography==2.8
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.10
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
